byte->int

+++ Runtime/Core/Process.cs
@@ -59,7 +59,7 @@ namespace JFramework.Net
             }
         }
 
-        public void ServerReceive(int clientId, ArraySegment<byte> segment, byte channel)
+        public void ServerReceive(int clientId, ArraySegment<byte> segment, int channel)
         {
             try
             {
